mod: package refactoring
