+++ src/main/java/org/primefaces/virusscan/VirusScanner.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.primefaces.virusscan.spi;
-
-import org.primefaces.virusscan.VirusException;
+package org.primefaces.virusscan;
 
 import java.io.InputStream;
