+++ src/main/resources/META-INF/services/org.primefaces.virusscan.VirusScanner
@@ -1,0 +1,1 @@
+org.primefaces.virusscan.impl.VirusTotalVirusScanner
