--- src/main/resources/META-INF/services/org.primefaces.virusscan.spi.VirusScanner
@@ -1,1 +1,0 @@
-org.primefaces.virusscan.spi.impl.VirusTotalVirusScanner
